fixed logic error
the college student would skip over the stop; i fixed this; then the
college student would stay on the stop; i fixed this. logic is
practically solid all the way through (except till the end, when the
player's tile number is greater than the amount of tiles we have coded.
let's finish graphics now and then we can finish that code after.